+++ apps/web/src/app/Teacher/question_bank/questions/page.tsx
@@ -4,7 +4,7 @@ import QuestionDetails from "../../../../components/QuestionDetails";
 import * as React from "react";
 import { useDisclosure } from "@chakra-ui/react";
 import { Heading, Box, Flex, Button, Spacer } from "@chakra-ui/react";
-import QuestionBankData from "../../../MockData/question_bank.json";
+import QuestionBankData from "../../../../MockData/question_bank.json";
 import QuestionBank_Table from "../../../../components/QuestionBank_Table";
 import Link from "next/link";
